Fixed Error:(6, 8) java: class algorithm is public, should be declared in a file named algorithm.java. Inside file PrimeFactorization, the name of public class was wrong.

+++ ciphers/ColumnarTranspositionCipher.java
@@ -117,7 +117,7 @@ public class ColumnarTranspositionCipher {
      * order to respect the Columnar Transposition Cipher Rule.
      */
     private static int numberOfRows(String word) {
-        if ((double) word.length() / keyword.length() > word.length() / keyword.length()) {
+        if (word.length() / keyword.length() > word.length() / keyword.length()) {
             return (word.length() / keyword.length()) + 1;
         } else {
             return word.length() / keyword.length();
